[refactor] Added all components to single file
[refactor] Added all components to single file and used ES6 syntax in checkAddedInCart function
