+++ src/components/index.js
@@ -1,0 +1,4 @@
+export { NavBar } from "./NavBar/NavBar";
+export { Footer } from "./Footer/Footer";
+export { ProductCard } from "./ProductCards/ProductCard";
+export { CartProduct } from "./ProductCards/CartProduct";
